+++ pkg/clients/clients.go
@@ -86,7 +86,7 @@ func (dcs *DynamicClientSet) ResourceClient(gvk schema.GroupVersionKind, namespa
 	}
 
 	// For namespaced Kinds, create a namespaced client. If no namespace is provided, use the "default" namespace.
-	namespaced, err := dcs.namespaced(gvk)
+	namespaced, err := dcs.NamespacedKind(gvk)
 	if err != nil {
 		return nil, err
 	}
@@ -132,10 +132,15 @@ func (dcs *DynamicClientSet) gvkForKind(kind kinds.Kind) (*schema.GroupVersionKi
 	return nil, fmt.Errorf("failed to find gvk for Kind: %q", kind)
 }
 
-func (dcs *DynamicClientSet) namespaced(gvk schema.GroupVersionKind) (bool, error) {
-	resourceList, err := dcs.DiscoveryClientCached.ServerResourcesForGroupVersion(gvk.GroupVersion().String())
+func (dcs *DynamicClientSet) NamespacedKind(gvk schema.GroupVersionKind) (bool, error) {
+	gv := gvk.GroupVersion().String()
+	if strings.Contains(gv, "core/v1") {
+		gv = "v1"
+	}
+
+	resourceList, err := dcs.DiscoveryClientCached.ServerResourcesForGroupVersion(gv)
 	if err != nil {
-		return false, err
+		return false, fmt.Errorf("failed to find server resources for GV: %q - %v", gv, err)
 	}
 
 	for _, resource := range resourceList.APIResources {