+++ CHANGELOG.md
@@ -15,6 +15,7 @@
 
 -   Move helm module into a directory (https://github.com/pulumi/pulumi-kubernetes/pull/512)
 -   Move yaml module into a directory (https://github.com/pulumi/pulumi-kubernetes/pull/513)
+-   Put all resources in specified provider namespace (https://github.com/pulumi/pulumi-kubernetes/pull/506)
 
 ### Bug fixes